Move of checkout Content pages into common
We will be generalizing these for birth / marriage. Moving without
making any other changes so that later diffs are smaller, as GitHub is
not great about diffing + file moving.
