+++ services-js/registry-certs/client/common/checkout/ShippingContent.tsx
@@ -7,16 +7,13 @@ import InputMask from 'react-input-mask';
 
 import PageLayout from '../../PageLayout';
 
-import { BreadcrumbNavLinks } from '../breadcrumbs';
+import { BreadcrumbNavLinks } from '../../death/breadcrumbs';
 
 import Cart from '../../store/DeathCertificateCart';
 import Order, { OrderInfo } from '../../models/Order';
-import { makeStateSelectOptions } from '../../common/form-elements';
+import { makeStateSelectOptions } from '../form-elements';
 
-import {
-  DeathOrderDetails,
-  OrderDetailsDropdown,
-} from '../../common/checkout/OrderDetails';
+import { DeathOrderDetails, OrderDetailsDropdown } from './OrderDetails';
 import makeShippingValidator from '../../../lib/validators/ShippingValidator';
 import { runInitialValidation } from './formik-util';
